Refactor/landing responsive (#92)
* refactor(landing): Update env file for mobile and improve landing UI

Added env_file to mobile service in docker-compose.yml for better environment management. Adjusted landing page hero text to be responsive and hid the navbar brand text on small screens for improved UI.

* refactor(landing-hero): Add typewriter effect to hero and refactor features

Introduced the 'react-simple-typewriter' package and updated the hero section to use a typewriter animation for language names. Refactored feature and screen refs with useCallback, streamlined GSAP animation logic, and cleaned up code for better readability and maintainability.

* refactor(landing-hero-headings): Refactor landing page hero heading layout

Replaces the single-line heading with a vertically stacked layout using a flex column. This improves readability and visual emphasis on the typewriter effect for language names.